lead-generator/lead-gen-code/requirements.txt
typer
feedparser
openai
httpx
psycopg2-binary
python-dotenv
rich
beautifulsoup4
tenacity
pgvector
